Shadow rework
Should be more acurate regarding shapes edges now (only polys for now)

Still has some issues, namely:
1. sometimes parts overlap (manifests as thin white lines)
2. sometimes parts are in the wrong order (?) (manifests as thin black lines)
3. at certain angles it is quite visible where the rays are at the outer parts of the circle when parts are blocked by a body

+++ src/util/math.cpp
@@ -7,7 +7,7 @@ namespace engine {
 	namespace util {
 		float b2AngleDeg(b2Vec2 from, b2Vec2 to) {
 			to-=from;
-			float a =  atan2(to.y, to.x) * 180 / fPI;
+			float a =  atan2f(to.y, to.x) * 180 / fPI;
 			if (a < 0) a += 360;
 			return a;
 		}
